- cleanup only proceeds if $ows_containers is non-empty - test.sh returns non-zero error code if any subtest failed - client picks node to publish change set to BEFORE appending change set locally

+++ clean.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 ows_containers=$(docker ps --filter "name=ows_nodejs_containernode1" -a -q)
-docker stop $ows_containers
-docker rm $ows_containers
-docker system prune -f
+
+if [[ "$ows_containers" != "" ]]; then
+    docker stop $ows_containers
+    docker rm $ows_containers
+    docker system prune -f
+fi
